Add blinking when 30s left on timer
Tmux does not support ANSI escape codes. The pomo command is run
on an interval, it would not show up long enough for the ANSI
blinking to take effect. Therefore, I found it better to have the
emoji blinking. When the time left is less than 30s I only show the
emoji on even numbers, creating a blinking effect.